requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
openai>=1.59.0
chromadb==0.5.7
python-dotenv==1.0.1
pydantic==2.9.2
pypdf>=5.0.0
